feat(ui): switch from gulp/bower to webpack/npm
Bower is deprecated and it is recommended to migrate to newer alternatives.

Gulp works well for small and straightforward projects but it becomes quickly
hard to maintain on a larger codebase. Also, Webpack modern features such as
tree shaking are very useful and allow us to save 1MB on CSS for example.

Overall, this makes the maintenance of the Console much easier. Everything is in
one place (`package.json`) with maximum flexibility:

- The current code just stays the way it is but already benefit from the new
framework (`import`, optional typescript/es7, etc.) - Any new code can be
written in TypeScript (with Babel) following latest standards with extensive
linting rules.

The README.md was updated too, I believe that the transition should be smooth +
the development experience was improved as a result of using webpack-dev-server.