fix(Console): display notifications for 2 minutes (rather than 15 secs)
Given that the row count is only shown in notifications, we need to display them
for longer so the users can have time to access the information.

This is a short term fix, the row count will be shown next to the refresh button
in the next iteration of the React refactor of the codebase.