chore(sql): removed 'position' attribute from Function interface
Position attribute is a compile-time requirement, which value we held on Function - a runtime component. This resulted in unnecessary instantiation of functions that are essentially the same in runtime but differ in their position in SQL. Removing position allows us to not create so many object instances and reduce memory footprint of the entire database instance.