feat(ilp): support timestamp fields
Extends ILP to support timestamp field type. Timestamp field values
are determined based on the `t` suffix similar to how it's done
with `i` suffix for long values.

The previous (undocumented) behavior was to interpret values like
`1465839830100399000t` of symbol type in case of UDP and of boolean
type (true) in TCP.