fix(sql): fix bind variables support in queries with indexed column in filter
A bind variable in a query with a filter over an indexed column was
completely removed from the filter by the intristics. That led to
the incorrect query result since the bind variable literal, e.g. '$1',
was used for the index lookup instead of the user provided bind
variable value.