add post-build script to remove `sourceMappingURL` from monaco assets
this way web devtools do not throw a warning about missing assets.

including sourceMaps would increase the deployable size by ~11MB which
is not acceptable, considering the files are only for development.

Due to the fact that web console consumes production monaco bundles (we do not transpile code ourselves), and because those bundles already include comment lines like `sourceMappingURL`, dev tools is looking for source maps, sending a request to a path (which is taxing serverside), sees 404 and yields a warning.

An easy fix is to doctor monaco bundles before consumption - simply
remove lines like this:

```
//# sourceMappingURL=../../../../min-maps/vs/base/worker/workerMain.js.map
```