fix `fromFetch` epic util to support wider range of `application/json` headers
it's possible for a server to respond with header like:
`Content-Type: application/json; charset=utf-8`

In this case, content is still json, but because of a strict `if`, it
doesn't get returned and instead a default value is used.

An example of such server is `webpack-dev-server`. With the strict `if`
check it is not possible to test values of
`assets/console-configuration.json`, because code falls back to default
values.