Remove dispatch queue
The code around the explicit dispatch queue was complicated and racy.
With this change the dispatching always happens on a single thread
which simplifies the code a lot while the end behavior of async offload
stays almost the same.

Also adds some tests for the concurrent producer and consumer sequences.