fix(sql): fix stack overflow in select and insert cache pooling
A stack overflow could occur in situation when a cached SELECT or
INSERT query container gets returned to the pool. Luckily, such
error would only happen with non-default settings for the pool
capacity.

The issue is fixed by splitting WeakAutoCloseableObjectPool into
two separate classes with clear purposes: WeakSelfReturningObjectPool
and WeakCloseableObjectPool. The first class is the one that is used
in PG Wire caches. It doesn't expose push() method and instead
assumes that the pooled objects return themselves to the pool as
a part of the close() method call.