chore(core): implement concurrent symbol index reader
New symbol index reader supports thread-safe cursors that
don't remap key and value memory on access. The assumption
is that the owner is responsible for calling of() method
before it starts accessing the cursors.

Symbol map reader class was migrated to the new index reader
as a step of the async offload symbol filters thread safety
fix (see #2107).