fix(pgwire): Fix initial connection properties parser
There was a bug where a value of each property was also used as a key for a property created out of thin air by a buggy pgwire parser.
So when a client sends a property with a value set to "user" then a buggy pgwire parser would create also a key "user" out of thin air with a value set as the next key.
Example:
2022-05-17T16:39:18.308689Z I i.q.c.p.PGConnectionContext property [name=user, value=admin] <-- this is a legit property
2022-05-17T16:39:18.308707Z I i.q.c.p.PGConnectionContext property [name=admin, value=database] <-- this is a property "invented" by a buggy pgwire parser
2022-05-17T16:39:18.308724Z I i.q.c.p.PGConnectionContext property [name=database, value=qdb] <-- a legit property set by a client
2022-05-17T16:39:18.308789Z I i.q.c.p.PGConnectionContext property [name=qdb, value=client_encoding] <-- again, a property created out of thin air

Thanks to @bziobrowski & @mariusgheorghies for help with analysis!