fix(sql): fix uncommon potential race in parallel filter execution
PageFrameSequence#await() method wasn't awaiting for the reduce counter
to be incremented to the number of dispatched tasks. It could be the case
(a very infrequent one) that one of the workers reduced a page frame,
then marked the task as done, but haven't incremented reduce counter yet.

The race manifested itself as the following AssertionError:
```
java.lang.AssertionError
	at io.questdb@6.3.1/io.questdb.cairo.sql.async.PageFrameSequence.await(PageFrameSequence.java:126)
        ...
```

This commit adds the missing await loop.