fix(sql): fix potential resource leak on update statement release
UpdateOperation release logic had a data race: startAsync() method
was called after the operation was published to a writer queue.
Hence, both close and closeWriter methods of UpdateOperation could
see different values of the executingAsync field and due to this
decide not to free the record cursor factory leaving its resources
leaked.