fix(sql): improve performance of certain SAMPLE BY queries
Closes #2146

Default FastMap size used in no fill SAMPLE BY queries was too large.
As a result the database was spending a lot of CPU time in
FastMap#clear() method doing nothing, but filling offsets memory with
-1 values.

This commit changes the default size for affected queries to a much
smaller value. It also adds a +1 shift for the offsets values, so
that the memory can be filled with 0 instead of -1 which should
perform on x86-64.