fix(sql): fix potential resource leak in non-keyed sample by with filter
The issue was twofold. First, both AsyncFiltered* cursors could throw
a filter-caused exception on of() call. This broke the close() contract
in all factories dealing with a base cursor. Second,
AbstractSampleByNotKeyedRecordCursorFactory was not closing the base
cursor properly.

As a result, exception thrown by the filter or early client disconnect
led to situation when the async filtered cursor wasn't properly closed.