chore(sql): introduce default offset in sample-by queries
fixes #2390

According to the grammar at https://questdb.io/docs/reference/sql/sample-by/#syntax
when you align “to calendar” then either “time zone” or “with offset” is mandatory.
So one could say it old behaviour is “working as designed” 🙈

I'd argue this behaviour is surprising. Even our own documentation uses samples
where timezone/offset is missing - it works when "align to calendar" is the last
token is a query. This change is making the timezone/offset optional in all cases
and when it's omitted then it defaults to `with offset '00:00'`.