Fixes #1
The function width_bucket() was not working correctly.

The function was not checking for the following conditions:

1. count must be greater than 0
2. low must be less than high
3. low must not be equal to high

The function was also not checking for NaN values.

The function was also not checking for runtime constants.

The function was also not checking for NaN values.

The function was also not checking for runtime constants.