use Long128 wrapper for Long128 and UUID columns
there are still cleanups needed, but tests are green locally.
Pushing to run tests on other architecture and to give reviewers
a chance to see how this version looks like. Personally I'm not crazy about it,
the stateful nature of wrappers makes things more complex and difficult
to reason about.