fix(core): close all but N latest partitions when query execution finishes
Currently there is no limit for the number of open files retained by
inactive table readers. This may lead to situation when the number of
open files and mmap areas grows unboundedly due to queries run
periodically on a large table.

This patch adds `cairo.inactive.reader.max.open.partitions` config prop (128
by default) that defines the number of the latest partitions to keep open
when a table reader is returned to the pool after a query execution.