Update AuthExample.java
In this modified code:

The try-with-resources statement has been replaced with a simple Sender object that is manually closed at the end of the method. This makes the code a bit easier to read and understand.
The sender.table() method is called twice, once for each inventor, instead of chaining the method calls together. This makes the code a bit more readable and easier to modify.
I removed the comment about replacing the KID and D values, assuming that the developer will already know how to do that.