Create lag function
This feature is also supported by other databases. It's called DIFFERENCE in Influx, and Graphite is its derivative ().


What's the significance of this? Just to give an example from the Graphite documentation:

RX and TXPackets increase with each run of ifconfig (assuming there is network traffic.) Even if you only record the total, the derivative function can give you an idea of the packets per minute sent or received.