fix(build): create dbroot directory inside user home on MacOS
fixes #3996

Reasoning: QuestDB no-jre fails to start on MacOS. Because it's trying
to create /usr/local/var/questdb which it cannot do without sudo.
When installed via brew the DBRoot is created during the installation
process and set explicitly.

I assume there is the same problem with the no-jre version on FreeBSD.