chore(test): fix a flaky pgwire cancellation test
fixes #3180

reasoning: if the query is completed before the cancelling
thread is scheduled for execution then the test fails. this can
happen on under-powered boxes (e.g. Windows CI) where a scheduler
has to shuffle a lot of tasks. this changeset makes the queries
slower thus the cancelling thread has a higher chance to cancel
the query.