fix(http): fix edge case error when importing a large CSV file
Reasoning:
This is the buggy line: `int diff = ((int) (nextOffset - offset)) - bytesToRead;`
`nextOffset` and `offset` are 48 bits numbers stored inside 64 bits long.

When subtracted from each other the result can produce be more than 32 bits.
When narrow-casting the result to `int` the higher bits are truncated.

If you are especially unlucky then the truncated `diff` value is 0 or 1
and the coalescing algo will coalesce lines which should not be coalesced.