Fix for testCaseInitsArgs and more tests.
This issue is caused by the form (select-group-by (select-choose (select))) where the inner select-choose is entity and elided during factory generation.

However, the corresponding model is not updated, so Async Group By codegen fails to find the correct WHERE condition.

A patch to search for a WHERE in the nested model helps to mitigate this issue.