Updating default shared worker values.
Changed lineTcpWriterWorkerCount to default to 0 instead of 1. This field is used to dedicate threads for ILP ingestion for non-WAL tables. This will increase default eligibility for AsyncGroupBy optimisation for 4 vCPUs.

Increase minimum shared.worker.count to 4, up from 2. This increases eligibility for AsyncGroupBy optimisation for small instances <4 vCPUs.

Updated testAllDefaults() to account for new default value.

Added test for shared workers to guarantee a minimum of 4.