fix deadlock on macos
on close() the closing thread writes into a pipe. this is meant
to unblock the blocked polling thread. in order for this to work
we have to register to receive read-ready events from the pipe.

this way the closing thread set the `closed` flag and writes into
the pipe. this notifies the poller, which checks the closed()
flag, realizes it's closed, exit waitForChange() and finally
unblock the latch the closing thread is waiting for.