feat(core): replace cairo snapshot instance ID with trigger file for snapshot recovery
Previously, snapshot recovery was driven by the Cairo Snapshot Instance ID
property, which had no default value and was generally confusing to use.

After this change, Cairo Snapshot Instance IDs are no longer used at all.
Snapshot recovery is now triggered by the existence of a trigger file in
the QuestDB root directory: If a file named root/_restore exists and a
snapshot is available, it will be restored on startup.