test(sql): fix flaky testServerMainCreateWalTableInVolume()
The original code was flooding the system with (parallel?) SAMPLE BY
queries until it returned the expected results.

There are 2 issues with this approach:
1. minor one: is the query occasionally returns wrong results then the test won't catch it
2. bigger one: it looks like the WAL Apply Job was starved due almost 5k SAMPLE BY queries running in the span of 5s.
  See the log: https://gist.githubusercontent.com/jerrinot/2efdc5d35f3d1cfbdad2fda78375db2a/raw/d51368ae264f262f1e64a6ad1cdce244d0ac3900/gistfile1.txt
  There are 5k queries, none of them returns the expected result. Only after the failure there is a log message about WAL being ejected.

We know the test populate the table atomically (INSERT ATOMIC INTO ...) thus
we can await for the tx to be applied and then simply assert the expected result.