test(core): fix flaky CheckpointTest.testSuspendResumeWalPurgeJob()
This PR https://github.com/questdb/questdb/pull/4869 introduced
the `StationaryMicrosClock` in `CheckpointTest.setUpStatic()`, which was
necessary since `testCheckpointStatus()` relied on frozen clocks.

However, other tests, such as `testSuspendResumeWalPurgeJob()`,
depend on the ability to manipulate time freely and fail with the `StationaryMicrosClock`.

This issue was mostly obscured because `CheckpointTest` resets the clock
to the default (non-StationaryMicrosClock) after each test case. Therefore,
`testSuspendResumeWalPurgeJob()` only fails when it runs as the first test
case in `CheckpointTest`.

When another test runs first, it resets the clock back to normal after completion,
and subsequent executions of `testSuspendResumeWalPurgeJob()` pass normally.

Fixes #4922

This fix removes the `StationaryMicrosClock`, and instead, `testCheckpointStatus()`
sets the time via `currentMicros`.