fix(ilp): ILP clients should not retry when inserting a bad UUID value
Previously, when an ILP client attempted to insert a non-UUID
string into a UUID column, QuestDB internals would
return an ImplicitCastingException.

This resulted in the HTTP server returning a HTTP 500
status code instead of the more appropriate HTTP 400.

Clients interpret an HTTP 500 status code as a transient
error and would retry. This change prevents a retry loop
by explicitly validating that the received value is an actual UUID.