feat(ilp): ILP client allows to cancel the current row
This is useful when you start a row and then realize
the row is invalid and cannot be sent to the QuestDB server.

This can occur when you cannot validate data upfront,
for example, when you receive them in a streaming fashion.

This API is implemented only for the HTTP transport, as
the TCP transport can flush at any time. The HTTP transport flushes
only after finishing a row.