fix inferring unspecified parameter types
when a client does not know a parameter type it sends
OID as 0. the old code interpreted it as string and
defined a string binding variable. this means the compiler
was forced to use the string type.

after this change we do not define unknown variables before
calling the compiler. this gives the compiler a chance to
infer the best possible type.