fix(sql): bug in SAMPLE BY queries after a query cache hit
In the old code `FillRangeRecordCursorFactory` would store
a mutable CharSequence with interval. This causes a problem
when the factory is stored in a query cache and in the next
execution the CharSequence instance is pointing to something else.

This PR fixes this by eagerly parsing the interval CharSequence
and storing just immutable values.