fix(sql): race parallel group by with many groups
When parallel group by sharding took place, it could happen that
a work-stealing thread would steal other thread's task and would
start executing it as if it was the task's owner.

The fix makes the per-worker synchronization mandatory in such
cases by checking the task's atom and comparing it with the
thread's one.