[PATCH] Avoid full scans for certain `LIMIT -N` scenarios (#4726)
This commit tries to push down negative `LIMIT` clauses in order to avoid a full table scan.

Previously, a `JOIN ... LIMIT -N` would trigger a full scan (going through the entire result set, and
only returning the last N rows). This commit tries to push down this requirement at the partition level
in cases where there are no filter constraints on the scan itself. Specifically, this is when the query
(or subquery) is sorted by the timestamp, in which case, a simple seeking into the relevant partition/row
is enough.

This optimization can also potentially be used in other cases where negative offsets are used, as the
current implementation utilizes a backwards scan and sort; but the sort step can also be eliminated
if there are no other constraints.