support for count(distinct col) and string_agg(distinct col)
This means that instead of using count_distinct(foo), it is now possible
to write count(distinct foo). Similarly, instead of string_distinct_agg(foo),
you can now use string_agg(distinct foo).

The motivation for this change is to improve compatibility with PostgreSQL.

Internally, the count() and string_agg() functions are rewritten to their
distinct versions when the distinct keyword is used.
Note for reviewers:

count(distinct) is still treated as a non-unique count of a column named
distinct. Since distinct is a keyword, we could (or should?) enforce that
it be enclosed in double quotes when used as a column name. This would cause
such queries to fail. However, this PR is more lenient and does not enforce
this, prioritizing backward compatibility.