fix(ilp): Java ILP client supports timestamps up to 9999-12-31 23:59:59.999999
Client no longer works with nanoseconds internally. Previously, even non-designated
timestamps were converted to nanos only to be converted back to microseconds. The highest
timestamp which can be represented in nanos in a 64-bit signed integer is 2262-04-11 23:47:16.854.
Anything above this would overflow. This change converts timestamps directly to microseconds,
preventing the overflow.

It also changes behavior for designated timestamps: they are now sent as microseconds instead
of nanos, since the server would store them in microseconds anyway. This means even the
designated timestamps now support a higher range.

The max timestamp that works reliably is 9999-12-31 23:59:59.999999 since our
timestamp format compiler assumes at most 4-digit years. If we remove this limitation,
we could support timestamps up to 294247-01-10T04:00:54.775807Z.

Last but not least: `WalAppender` now checks if the micros value is non-negative and throws
an appropriate error when it's not. Without this check, an attempt to store a negative
designated timestamp resulted in a CairoException which was propagated as an internal error.
Internal errors are represented as HTTP 500 and this makes clients retry, even when
it's futile in this case. An explicit check with `LineProtocolException` results in
HTTP 400 which is NOT retried by clients.