re-introduce EqTimestampStrConstantFunction
also a behaviour change: all version of "interval search" are fail-fast.
previously, constant versions were forgiving - when the interval couldn't be parsed the exception was ignored and in() never matched. however, other versions were most strict: a failure to parse the interval caused an error. this change unifies both on the fail-fast behaviour