fix(build): load native libs from distribution lib dir, not /tmp
Fixes #570

On security-hardened systems with a `noexec` /tmp mount,
QuestDB fails to start because it cannot execute native libraries copied there.

This commit alters the build process to place native libraries for Linux and
macOS into a `lib/` directory within the distribution. The server now loads
them directly from this location, avoiding the `/tmp` directory entirely.

This increases the distribution size slightly due to library
duplication but this is a good trade-off for increased compatibility.