perf(http): garbage-free HTTP header parser
This change reduces the allocation rate when parsing HTTP headers.
Previously, header parser would carefully pool HTTP headers,
only to store them in a map which implicitly created a defensive
copy of all keys.

This change extends the map with a method to store keys
without creating a defensive copy.