ci(build): delete dangling files in /tmp on Hetzner
It looks like tests occasionally keep dangling files in /tmp.
It is likely caused when a test is cancelled. This is invisible
on Azure-hosted runners since they always start with a clean table.

Dangling files/dirs accumulate on self-hosted runners. This change removes
artifacts known to accumulate before starting tests.